Use i128

+++ 2020/rust/day10/src/main.rs
@@ -205,7 +205,7 @@ fn main() {
         let mut jolts: Vec<usize> = lines.map(|line| line.unwrap().parse::<usize>().unwrap()).collect();
         jolts.sort();
         jolts.push(jolts.last().unwrap() + 3);
-        let mut paths: HashMap<i32, i32> = (-3_i32..*jolts.last().unwrap() as i32 + 3).map(|jolt| (jolt.clone(), 0)).collect();
+        let mut paths: HashMap<i32, i128> = (-3_i32..*jolts.last().unwrap() as i32 + 3).map(|jolt| (jolt.clone(), 0)).collect();
         // paths[&0] = 1;
         // trait `IndexMut` is required to modify indexed content, but it is not implemented for `std::collections::HashMap<usize, usize>`
         // https://stackoverflow.com/questions/30414424/how-can-i-update-a-value-in-a-mutable-hashmap
